refactor(T10132): decompose nexus.ts renderers into core/render/nexus/{graph,contracts,audit} (B7)
Migrates 34 Nexus renderers (1055 LOC) from packages/cleo/src/cli/renderers/nexus.ts
to packages/core/src/render/nexus/. Family layout:
  - graph/  — 22 renderers: status, setup, clusters, flows, context, impact,
              analyze, query, route-map, full-context, hot-paths, hot-nodes,
              search-code, why, impact-full, refresh-bridge, shape-check,
              projects-{list,register,remove,scan,clean,clean-preview}
  - contracts/ — 8 renderers: contractsSync/Show/LinkTasks, conduitScan,
              brainAnchors, taskFootprint, taskSymbols, wiki
  - audit/  — 3 renderers: diff, coldSymbols, export

Each renderer is migrated verbatim — zero behaviour change. The cleo dispatcher
now imports the renderers from @cleocode/core; nexus.ts is deleted. Pure-renderer
snapshot/assertion tests moved to packages/core/src/render/nexus/__tests__/
renderers.test.ts (15 tests, all green); cliOutput dispatcher integration tests
stay in packages/cleo/ (3 tests, all green). 8 cleo renderer test files (186
tests) and 3 core render test files (55 tests) all pass unchanged.

Inline tree/table duplication is left intact with TODO(T10128 B3 primitive ready)
markers — B3 (T10128) animation primitives just landed; a follow-up PR (or
T10137 B12 closeout) can collapse the inline tables into the B3 primitives.

cli-error-signature.test.ts mock for @cleocode/core was widened to include the
nexus renderer stubs the dispatcher now imports.

Resolves AGENTS.md Package-Boundary Check violation — CLI commands package was
hosting business logic that belongs in core.

Subtasks delivered: T10150 (graph), T10151 (contracts), T10152 (audit).

Closes T10132
Subtasks: T10150, T10151, T10152
Epic: T10114
ADR: adr-077-human-render-contract
