+++ packages/cleo/src/cli/renderers/index.ts
@@ -32,6 +32,41 @@ import {
   formatSuccess,
   metaFooter,
   pagerFooter,
+  // Nexus renderers — migrated to @cleocode/core/render/nexus per ADR-077 (T10132 / B7).
+  renderNexusAnalyze,
+  renderNexusBrainAnchors,
+  renderNexusClusters,
+  renderNexusColdSymbols,
+  renderNexusConduitScan,
+  renderNexusContext as renderNexusContextResult,
+  renderNexusContractsLinkTasks,
+  renderNexusContractsShow,
+  renderNexusContractsSync,
+  renderNexusDiff,
+  renderNexusExport,
+  renderNexusFlows,
+  renderNexusFullContext,
+  renderNexusHotNodes,
+  renderNexusHotPaths,
+  renderNexusImpact,
+  renderNexusImpactFull,
+  renderNexusProjectsClean,
+  renderNexusProjectsCleanPreview,
+  renderNexusProjectsList,
+  renderNexusProjectsRegister,
+  renderNexusProjectsRemove,
+  renderNexusProjectsScan,
+  renderNexusQuery,
+  renderNexusRefreshBridge,
+  renderNexusRouteMap,
+  renderNexusSearchCode,
+  renderNexusSetup,
+  renderNexusShapeCheck,
+  renderNexusStatus,
+  renderNexusTaskFootprint,
+  renderNexusTaskSymbols,
+  renderNexusWhy,
+  renderNexusWiki,
 } from '@cleocode/core';
 import type { CliEnvelope, CliMeta, Warning } from '@cleocode/lafs';
 import { applyFieldFilter, extractFieldFromResult } from '@cleocode/lafs';
@@ -69,43 +104,6 @@ import {
 export type { RenderWavesMode, RenderWavesOptions } from './system.js';
 export { renderWaves };
 
-// Nexus renderers (T1720)
-import {
-  renderNexusAnalyze,
-  renderNexusBrainAnchors,
-  renderNexusClusters,
-  renderNexusColdSymbols,
-  renderNexusConduitScan,
-  renderNexusContext as renderNexusContextResult,
-  renderNexusContractsLinkTasks,
-  renderNexusContractsShow,
-  renderNexusContractsSync,
-  renderNexusDiff,
-  renderNexusExport,
-  renderNexusFlows,
-  renderNexusFullContext,
-  renderNexusHotNodes,
-  renderNexusHotPaths,
-  renderNexusImpact,
-  renderNexusImpactFull,
-  renderNexusProjectsClean,
-  renderNexusProjectsCleanPreview,
-  renderNexusProjectsList,
-  renderNexusProjectsRegister,
-  renderNexusProjectsRemove,
-  renderNexusProjectsScan,
-  renderNexusQuery,
-  renderNexusRefreshBridge,
-  renderNexusRouteMap,
-  renderNexusSearchCode,
-  renderNexusSetup,
-  renderNexusShapeCheck,
-  renderNexusStatus,
-  renderNexusTaskFootprint,
-  renderNexusTaskSymbols,
-  renderNexusWhy,
-  renderNexusWiki,
-} from './nexus.js';
 // Task renderers
 import {
   renderAdd,
