+++ packages/core/src/render/index.ts
@@ -13,6 +13,7 @@
 export * from './ansi.js';
 export * from './fallback.js';
 export * from './helpers.js';
+export * from './nexus/index.js';
 export * from './registry.js';
 export * from './render-envelope.js';
 export * from './types.js';
